requirements.txt
absl-py
albumentations
asn1crypto
astor
attrs
cachetools
certifi
cffi
chardet
cryptography
cycler
gast
gdown
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
joblib
Keras-Applications
Keras-Preprocessing
keyboard
kiwisolver
Markdown
matplotlib
more-itertools
natsort
nltk
numpy
oauthlib
openai
opencv-python
opt-einsum
packaging
pandas
Pillow
plotly
pluggy
portalocker
protobuf
py
pyasn1
pyasn1-modules
pycosat
pycparser
pyOpenSSL
pyparsing
PySocks
pytest
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
retrying
rsa
ruamel-yaml
sacrebleu
scikit-learn
scipy
sentencepiece
six
tensorboard
tensorflow
tensorflow-estimator
termcolor
torchtext==0.6.0
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchaudio==2.2.2+cu118
--extra-index-url https://download.pytorch.org/whl/cu118
tqdm
typing
urllib3
warmup-scheduler
wandb
wcwidth
Werkzeug
wget
wrapt
xmltodict
zipp
